+++ core-api/src/main/java/Hongik_SafeMap_Server/domain/admin/disaster_review/dto/response/AdminReportPageResponse.java
@@ -1,4 +1,4 @@
-package Hongik_SafeMap_Server.domain.admin.disaster_review;
+package Hongik_SafeMap_Server.domain.admin.disaster_review.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
@@ -6,7 +6,7 @@ import java.util.List;
 
 public record AdminReportPageResponse(
         @Schema(description = "제보 목록")
-        List<AdminReportListResponse> reports,
+        List<AdminReportResponse> reports,
         
         @Schema(description = "현재 페이지 번호 (0부터 시작)", example = "0")
         int currentPage,
